move and rename hitpoint.h/.cpp to slg/textures/hitpoint/hitpointcolor.h/.cpp
