+++ include/slg/textures/hitpoint/hitpointcolor.h
