+++ src/slg/scene/parsetextures.cpp
@@ -50,8 +50,8 @@
 #include "slg/textures/fresnel/fresnelpreset.h"
 #include "slg/textures/fresnel/fresnelsopra.h"
 #include "slg/textures/fresnel/fresneltexture.h"
+#include "slg/textures/hitpoint/hitpointcolor.h"
 #include "slg/textures/hitpoint/shadingnormal.h"
-#include "slg/textures/hitpoint.h"
 #include "slg/textures/hsv.h"
 #include "slg/textures/imagemaptex.h"
 #include "slg/textures/irregulardata.h"
